usb-device: Ignore most some of the Mbed OS /features/ to reduce the compile time.
It took over 5 minutes to do a complete compile and link of this very simple application.

The Mbed CLI likes to compile every source file it find even though most are being used.

'.mbedignore' can help with this and reduce the compile time to just over a minute. There are probably more savings that could be made but I'm keeping it simple at the moment.