usb-device: Corrected the description of where to find the Windows distribution of OpenOCD.
I found the correct/latest website for the Windows distribution of OpenOCD.

It described where I should have installed OpenOCD so I moved it.

Now the path doesn't contain spaces I removed the quotation marks.

It also seems it is unnecessary to specify the full path for the configuration files.