usb-device: Changed the /interface subclass/ to 0xff.
It is not clear to me that the interface subclass of a vendor specific class is important.

But, I saw some examples where it was done like this, i.e. 0xff, 0xff, 0xff, so I thought I'd do the same.

The STM debug interface to the Nucleo looks like this.