usb-device: Build Mbed OS debug profile.
I've been using gdb and OpenOCD to run and debug. Given that the app is so small may as well build with the debug options.

I've found that OpenOCD only connects to the target correctly when I've got the reset button pressed. I'm not sure what this is about but it's not a problem at the moment.