usb-device: Made 'configuration_descriptor' a member of the class instead of being static.
The lifetime of 'configuration_descriptor' is important, it is returned to the stack so it can't just be /automatic/.

It worked as a static for experiments I've done so far but to add end points its value needs calculating at run time, i.e. the endpoint address is not known at compile time.

I've copied what is done in 'USBCDC'. I've no doubt that there are other options but it will do.