usb-device: Perform a bulk in transfer.
/In/ transfers move data from the device to the host.

The device starts the transfer as soon as the USB device is configured.

Nothing happens until the host does its bit. All USB transfers are instigated by the host, so it is the host that actually does the transfer.

If the host doesn't do the transfer the device just waits.

This example is simplistic and inflexible, i.e. the length is fixed and it has to happen, but it's a start.