usb-device: Create custom target for the 32F723EDISCOVERY Discovery Kit.
The STM32F723ZEI6 has an on-chip Hi-speed PHY which clearly makes it a good choice for experimenting with High Speed USB.

There's a discovery kit wit a STM32F723IE...

https://www.st.com/content/st_com/en/products/evaluation-tools/product-evaluation-tools/mcu-mpu-eval-tools/stm32-mcu-mpu-eval-tools/stm32-discovery-kits/32f723ediscovery.html

But, unfortunately, Mbed OS doesn't include this discovery kit as a target.

As a special treat I'm going to try creating my own Mbed OS target.

The instructions are here...

https://os.mbed.com/docs/mbed-os/v6.1/program-setup/adding-and-configuring-targets.html

I thought I create a branch a do it in small steps, I've got a feeling it will be a while until I've got it working.