usb-device: Added linker script for 'DISCO_F723IE' target.
The next error came because the CLI tools looked early for a linker script and failed to find one.

Here it is!

The directory structure is modelled on what's in 'mbed-os/targets/TARGET_STM'.

I copied 'mbed-os/targets/TARGET_STM/TARGET_STM32F7/TARGET_STM32F746xG/device/TOOLCHAIN_GCC_ARM/STM32F746xG.ld', renamed it and changed the flash size.

I expecting, hoping, that everything else will be the same.

Adding the linker script wasn't enough on its own. It was also necessary to add the labels which must be related to the directory structure in some way.