usb-device: Use mbed-os-6.1.0 version of 'rtos::ThisThread::sleep_for'.
Specifying the delay as a /chrono/ literal removes warnings about the old version of 'sleep_for' being deprecated.

Annoyingly I knew I was supposed to be using this version of 'sleep_for' and tried before and failed.

I thought I was using mbed-os-6.1.0 but now I've realised I wasn't it is not a surprise that it didn't work.