usb-device: Configure the HSE frequency for the 'DISCO_F723IE' target.
'HSE_VALUE' is used in the STM32F7 HAL, see stm32f7xx_hal_conf.h, and the default value is 8 MHz.

Hence it needs to be set to 25 MHz for the 32F723EDISCOVERY.

With this set correctly usb-device starts up and the serial output can be seen. I haven't yet tried the USB FS.