usb-device: Conditionally compile all the code that is dependant on the Mbed-OS USB support.
I want to try going direct to the STM32 USB HAL because I think it will provide more flexibility and help me understand USB better.

I thought about creating a new repo for this, or just doing it on ^master^. In the end I created a branch which feels like the middle ground.