usb-device: Style consistency.
I copied 'OTG_HS_IRQHandler' from a C file.

At the time I think I thought the 'void' might be relevant because 'OTG_HS_IRQHandler' is declared 'extern "C"'. I've changed my mind and decided it's not needed and not in keeping with a C++ file.

Also, I failed to make the brace position consistent with the rest of the file.