usb-device: Define USB constants using scoped enums.
I think it is uncontroversial to use enums to define a list of related values.

Whether or not to make them scoped enums is a bit more controversial. Because they are /scoped/ there's no implicit conversion to the underlying type and 'to_underlying' is required for the bit manipulation.

Given these constants are only for bit manipulation and they are there's am outer namespace perhaps /scoping/ them is overkill!