usb-device: Refactored to create a separate enum for the 'bmRequestType' masks.
This time I haven't used a /scoped/ enum so that there is an implicit cast to an integer type.

I've done this mainly because having the mask in the other enums meant I couldn't use them in switch statements without getting the "enumeration value 'mask' not handled in switch" warning.