usb-device: Keep track of the USB device state.
The USB spec describes the various states a device can be in.

Up until now it hasn't affected what I've been doing but I think it is needed to move on.

What I've done is tested. The only test I've done is to reset the STM32 and move from 'default', through 'addressed' to 'configured'.