usb-device: Handle the vendor control in transfer requested by usb-host.
This is the second of the test transfers done by usb-host. To run this test I've got the bulk test transfers commented out.

What's already in 'HAL_PCD_DataInStageCallback' is important to the success of the transfer. The must ack sent by the host in the status stage of the sequence must be recieved.