usb-device: Handle the EP1 bulk out transfer requested by usb-host.
To test this the bulk in part of 'usb-host' was commented out.

I checked the data was received by breaking in 'HAL_PCD_DataOutStageCallback' and having a look at 'ep1_receive_buffer'. Nothing useful is done with the data that's for another day.

The details of the transfer can be found in 'hpcd->OUT_ep[1]'.