usb-device: Handle endpoint clear request.
I saw this request after attempting to enable the DMA feature. I don't think it is directly related to DMA, that was just a coincidence.

This is a simplified version of what I found in 'STM32Cube_FW_F7_V1.16.0/Middlewares/ST/STM32_USB_Host_Library/Core/Src/usbh_ctlreq.c'.

It didn't make the DMA feature work but I believe it is the correct thing to do.