usb-device: Specify the bulk transfer length independently from the MPS.
For no specific reason I had it in my mind that the transfer length needed to be the same as the endpoints maximum packet size.

This isn't true, the HAL and libusb are both capable of splitting a bulk transfer across multiple packets.

I think I understand how the HAL does this for IN transfers:
    * 'HAL_PCD_EP_Transmit' configures the transfer and enables the endpoint, it doesn't copy any data into the TX FIFO
    * The transfer configuration includes the total transfer length and the number of packets required
    * A 'Transmit FIFO empty' interrupt is generated when the endpoint is enabled
    * 'HAL_PCD_IRQHandler' takes the interrupt and decodes it
    * 'HAL_PCD_IRQHandler' calls 'PCD_WriteEmptyTxFifo' which calls 'USB_WritePacket'
    * 'PCD_WriteEmptyTxFifo' ensures the TX FIFO is full (or at least contains as many whole packets as it can)

I'm a bit vague on the precise operation of 'PCD_WriteEmptyTxFifo'. I have a suspicion that it will only put whole packets in the TX FIFO which means, if the TX FIFO is not a multiple of the MPS, it will not be filled completely.

In this commit the transfer size is still the same as the MPS, that's just because it is easier to test and integrate with 'usb-host'.