usb-device: Use 'usb_device::bulk_transfer_length' in 'USBDevice' build.
'main.cpp' now uses 'usb_device::bulk_transfer_length' as was intended.

The only problem is that 'USBDevice::write_start' asserts if the transfer size is bigger than the MPS.

This is a big problem in comparison to the HAL and restricts the throughput.

I've added an assert to 'EvkUSBDevice::bulk_in_transfer' and 'EvkUSBDevice::bulk_out_transfer' because it felt that a good way to document what's going on.