usb-host: No longer check the bulk IN data.
I've been running this test a great deal now and the bulk IN data is always what is expected.

Now that the transfer size is bigger the test is affecting the result. Removing it increased the throughput by 9 Mbit/s in the test I've just done.

At some point I might want the 'usb-host' to do something with the data, which might be akin to this test, but for now lets just measure the maximum throughput.