usb-device: Enable the USB HS DMA.
When transferring large packets it makes sense to use the DMA. For each bulk packet there's 512 bytes to move to/from the FIFO from/to system RAM.

This didn't affect the throughput in my test very much but then I'm not trying to do anything else with the CPU.

The DMA should just work but it turns out it is a bit fussy about the FIFO sizes and destination buffers.

I had to make the destination buffer 64 bytes and the Tx FIFO 0x100 words. Hay hoe!