spi-master: Create SPI master for testing the SPI slave of 'usb-device'.
I want 'usb-device' to something more useful. The plan is to make it take data in on a SPI interface and then send it out on the USB interface.

Before I started working on 'usb-device' I thought it would be good to have a SPI master to generate the data.

I contemplated doing a ^SPI loopback^ on 'usb-device' but decided using another board would be easier and less confusing.

It just so happens I have a NUCLEO-F767ZI sitting on my desk next to the STM32F723E-DISCO so I'm going to try using it as my source of SPI data.

I've copied the bare bones of am Mbed OS project from 'usb-device', it compiles and runs but doesn't do anything yet.