spi-master: Use DMA to move data to the SPI peripheral.
This means the SPI transactions come back-to-back, well almost.

I had to raise the priority of the 'show_running' thread to keep the LED flashing. With such a short transaction the 'spi_tx' only waits for 5us and obviously doesn't allow the lower priority thread to be scheduled.

NSS is only deasserted (set) for 50ns, that must be how long it takes to loop and call 'LL_GPIO_SetOutputPin' and 'LL_GPIO_ResetOutputPin'.

After SCLK stops it takes 5us to deassert (set) NSS which must be the time required to generate the interrupt, service it and set the thread flag.

This is longer than what I saw with the blocking function, NSS got deasserted in 2us.

Is any of this important? Probably not!

The DMA code was based on what I found in 'STM32Cube_FW_F7_V1.16.0/Projects/STM32F767ZI-Nucleo/Examples/SPI/SPI_FullDuplex_ComDMA'.