spi-master: Create a continuous SPI transfer.
I not sure if 'SPI transfer' is the correct term but I wanted SCLK to clock continuously and this is what I done.

It might be more useful if the data changed but I'm going to leave this for now and think about the the other end i.e. 'usb-device'.

It can stopped, and started again, using the blue user button.