usb-device: Removed the Mbed OS 'USBDevice' example code.
I now believe 'USBDevice' restricts USB performance and optimum performance requires going direct to the HAL.

The downside of the HAL is that it is target specific but portability isn't a requirmenet of this project.

I want to add some more stuff to 'usb-device' and thought it would be good to have a clear out first.