usb-device: Implement SPI data receiver.
The plan is to ingest data from SPI and forward using USB.

Ingesting some data is the first, small, step.

If the STM32F723-DISCO is connected to the NUCLEO-F767ZI the disco's red LED flashes.

I don't yet do anything with the data and I have a suspicion the DMA needs to be more complicated, i.e. multiple buffers.