usb-device: Consistent use of 'extern "C"'.
An explicit 'extern "C"' isn't necessary for a function definition if there is already an appropriate declaration.

The definition of 'OTG_HS_IRQHandler' required an 'extern "C"' because its weak definition is in 'startup_stm32f723xx.s' and there's no declaration.

The HAL PCD functions are weakly defined in 'stm32f7xx_hal_pcd.c' but importantly there is a declaration in 'stm32f7xx_hal_pcd.h'. The declaration means the 'extern "C"' is needed for the definition.

But, in 'spi-rx.cpp' by included the 'extern "C"' with the function definitions.