usb-device: Moved callbacks into namespace for translation unit.
It has just occurred to me that 'extern "C"' functions can be in the namespace and therefore the local variables don't need qualifying.

Is this a /normal/ thing to do? I've no idea but it works for me for now.

The alternative was to refactor the body of these functions into local functions but that felt like a bit of a waste.