usb-host: Copied the data check from 'usb-device'.
'usb-device' does a quick check that the SPI rx data is what we expect. It is, some of the time.

More interesting is whether or not the data is still what we expect by the time it has been moved to the host via USB.

Copying the test from 'usb-device' gives a superficial indication that the data is still what we expect.