usb-device: Consistent use of 'std::chrono_literals'.
Only import the symbol that is actually being used. Limiting the number of imported symbols feels like a C++ idiom to me but in this case I'm probably /going over the top/.

Import the symbols at the beginning of the enclosing scope.

I'm just being consistent with what I've just done in 'spi-master'.