spi-device: Added function for checking the overflow buffers.
I'm still trying to understand the SPI RX data and looking at the overflow buffers is easiest for now.

If used the SPI should be off before pressing the button otherwise 'find_expected_rx_pattern' and the DMA could both be writing to the overflow buffers.

This is just temporary code. I look forward to reviewing it in 5 years! ;-)