spi-master: Don't start sending SPI data until requested.
Use the blue push button to start sending data.

I see incorrect data sometimes and I think that might be because configuring and enabling the SPI slave while the master is transmitting is not allowed.

In the reference manual it says:
    28.5.8 Procedure for enabling SPI
    It is recommended to enable the SPI slave before the master sends the clock. If not,
    undesired data transmission might occur. ... The SCK signal must be settled at an idle state level
    corresponding to the selected polarity before the SPI slave is enabled.

I think this means 'usb-device' must be allowed to boot before the 'spi-master' starts transmitting.

Hence it doesn't make much sense for 'spi-master' to start transmitting immediately.