usb-device: spi-master: Added debug trace library.
I've added the debug trace library because I think it is a useful library although I'm not quite sure what I'm going to do with it yet.

Debug trace is not the same as expected output hence why I haven't added the version greetings to the trace.

I think calls to standard I/O output, i.e. 'puts' and 'printf', should probably also be protected by the serial out mutex. But, I'm going to leave that for now.

I've just realised that 'mbed-client-cli' might be better than what I've done and it uses the trace library, which is what instigated this change.