spi-master: No longer use the button to toggle the DMA.
I think it makes things simpler to just use the CLI.

No need to worry about 'toggle-dma' being called from multiple places at the same time and whether or not to call it via the event queue.

I think it is safe to run it on the CLI thread now it isn't called by the event queue.